requirements.txt
python-dotenv==1.0.1
smolagents[openai]==1.11.0
openai-agents==0.0.4
opentelemetry-instrumentation-openai==0.38.12
